requirements.txt
flask
flask-cors
google-generativeai
gunicorn
google-genai
dotenv
flask-caching
